Initial Commit
Proof-of-Concept letting you create a "study" which generates a unique URL to itself and lets you specify the amount of rounds, and then switch back and forth between them without refreshing through arrow controls.

All the rounds are just placeholders with simply the round number and a random element rn; plan is to fetch each round's content from the Model in my little component for loop in frontend\src\surveynewtest.js.