Public Experiment fixes
Fixed inkBlots in the Rorschach, fixed there being 2 Tip Rounds, generated the public experiments, and commented their code out. This db.sqlite3 should have all the stuff generated with no other data like testing accounts or experiments in it, i.e it should be good to send to Klaas.

Should be. Someone else please test it too to be safe.